refactor: Adding AoeffModel interface
