--- src/lib/common/interfaces/AoeffModel.ts
@@ -1,12 +1,0 @@
-interface AoeffXmlModel {
-  identifier: string;
-  xmlns: string;
-  'xmlns:xsi': string;
-  'xsi:schemaLocation': string;
-}
-
-export interface AoeffModel {
-  model: {
-    $: AoeffXmlModel;
-  };
-}
